Package reorganization (#1229)
* Moving top-level imports around

* AB import cleanup

* Restore module logging

* Splitting the controllers

* Making controllers stand-alone

* Fixed plugincontroller closing

* Fixed tests and older python issue

* Removed a testing property

* Update some formatting

* Run tests on minor branch

---------
