+++ .github/workflows/main.yaml
@@ -3,6 +3,7 @@ on:
   pull_request:
     branches:
       - master
+      - minor
   push:
     branches:
       - master
