+++ activity_browser/application.py
@@ -1,37 +1,52 @@
 # -*- coding: utf-8 -*-
-import logging
+import os
 
-from .controllers import controllers
-from .layouts import MainWindow
-from .logger import ABHandler
-from .info import __version__ as ab_version
+from PySide2.QtWidgets import QApplication
+from PySide2.QtCore import QSysInfo, Qt, QCoreApplication, QObject
 
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
+from activity_browser import log
 
+class ABApplication(QApplication):
+    _main_window = None
+    _controllers = None
 
-class Application(object):
-    def __init__(self):
-        self.main_window = MainWindow(self)
+    @property
+    def main_window(self) -> QObject:
+        """Returns the main_window widget of the Activity Browser"""
+        if self._main_window: return self._main_window
+        raise Exception("main_window not yet initialized, did you try to access it during startup?")
 
-        # Provide the log file path, so it is available for the Activity-Browser
-        # debug window/text box
-        log.info(f'The Activity Browser log file can be found at {log.log_file_path()}')
-        # Provide AB version for user debugging and that have that info stored in the logs
-        log.info(f'Activity Browser version: {ab_version}')
-        # Instantiate all the controllers.
-        # -> Ensure all controller instances have access to the MainWindow
-        # object, this propagates the 'AB' icon to all controller-handled
-        # dialogs and wizards.
-        for attr, controller in controllers.items():
-            setattr(self, attr, controller(self.main_window))
-
+    @main_window.setter
+    def main_window(self, widget):
+        self._main_window = widget
+    
     def show(self):
         self.main_window.showMaximized()
 
     def close(self):
-        self.plugin_controller.close_plugins()
-        self.main_window.close()
+        for child in self.children():
+            if hasattr(child, "close"):
+                child.close()
 
     def deleteLater(self):
         self.main_window.deleteLater()
+
+
+if QSysInfo.productType() == "osx":
+    # https://bugreports.qt.io/browse/QTBUG-87014
+    # https://bugreports.qt.io/browse/QTBUG-85546
+    # https://github.com/mapeditor/tiled/issues/2845
+    # https://doc.qt.io/qt-5/qoperatingsystemversion.html#MacOSBigSur-var
+    supported = {'10.10', '10.11', '10.12', '10.13', '10.14', '10.15'}
+    if QSysInfo.productVersion() not in supported:
+        os.environ["QT_MAC_WANTS_LAYER"] = "1"
+        os.environ["QTWEBENGINE_CHROMIUM_FLAGS"] = "--disable-gpu"
+        log.info("Info: GPU hardware acceleration disabled")
+
+if QSysInfo.productType() in ["arch", "nixos"]:
+    os.environ["QTWEBENGINE_CHROMIUM_FLAGS"] = "--no-sandbox"
+    log.info("Info: QtWebEngine sandbox disabled")
+
+QCoreApplication.setAttribute(Qt.AA_ShareOpenGLContexts, True)
+
+application = ABApplication()
