+++ activity_browser/bwutils/commontasks.py
@@ -9,11 +9,7 @@ from bw2data import databases
 from bw2data.proxies import ActivityProxyBase
 from bw2data.project import ProjectDataset, SubstitutableDatabase
 
-import logging
-from activity_browser.logger import ABHandler
-
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
+from activity_browser import log
 
 """
 bwutils is a collection of methods that build upon brightway2 and are generic enough to provide here so that we avoid
