+++ activity_browser/bwutils/superstructure/excel.py
@@ -6,13 +6,8 @@ from typing import List, Union
 import openpyxl
 import pandas as pd
 
+from activity_browser import log
 from .utils import SUPERSTRUCTURE
-
-import logging
-from activity_browser.logger import ABHandler
-
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
 
 
 def convert_tuple_str(x):
