+++ activity_browser/bwutils/superstructure/manager.py
@@ -1,30 +1,23 @@
 # -*- coding: utf-8 -*-
 import itertools
-from typing import List
+from typing import Union, Optional, List
+
 import numpy as np
 import pandas as pd
+import brightway2 as bw
 from pandas.api.types import is_numeric_dtype
 from PySide2.QtWidgets import QApplication, QPushButton
 from PySide2.QtCore import Qt
-from typing import Union, Optional
 
-import brightway2 as bw
-
+from activity_browser import log
 from .activities import fill_df_keys_with_fields, get_activities_from_keys
 from .dataframe import scenario_columns
 from .utils import guess_flow_type, SUPERSTRUCTURE, _time_it_
-
 from .file_dialogs import ABPopup
 from ..errors import (CriticalScenarioExtensionError, ScenarioExchangeNotFoundError,
                       ImportCanceledError, ScenarioExchangeDataNotFoundError,
                         UnalignableScenarioColumnsWarning, ScenarioExchangeDataNonNumericError
                       )
-
-import logging
-from activity_browser.logger import ABHandler
-
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
 
 
 EXCHANGE_KEYS = pd.Index(["from key", "to key"])
