+++ activity_browser/bwutils/superstructure/utils.py
@@ -1,13 +1,10 @@
 # -*- coding: utf-8 -*-
-import brightway2 as bw
-import pandas as pd
 import time
 
-import logging
-from activity_browser.logger import ABHandler
+import brightway2 as bw
+import pandas as pd
 
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
+from activity_browser import log
 
 # Different kinds of indexes, to allow for quick selection of data from
 # the Superstructure DataFrame.
@@ -62,6 +59,7 @@ def guess_flow_type(row: pd.Series) -> str:
         return "technosphere"
 
 def _time_it_(func):
+    #TODO rename to non_protected name
     """
     For use as a wrapper to time the execution of functions using the python time library
     """
