+++ activity_browser/controllers/plugin.py
@@ -1,28 +1,17 @@
 # -*- coding: utf-8 -*-
-import sys
 import importlib.util
-import traceback
 from pkgutil import iter_modules
-from shutil import rmtree
 
 from PySide2.QtCore import QObject, Slot
 
+from activity_browser import log, signals, project_settings, ab_settings, application
 from ..ui.wizards.plugins_manager_wizard import PluginsManagerWizard
-from ..signals import signals
-from ..settings import project_settings, ab_settings
-
-import logging
-from activity_browser.logger import ABHandler
-
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
 
 
 class PluginController(QObject):
 
     def __init__(self, parent=None):
         super().__init__(parent)
-        self.window = parent
         self.connect_signals()
         # Shortcut to ab_settings plugins list
         self.plugins = ab_settings.plugins
@@ -35,7 +24,7 @@ class PluginController(QObject):
 
     @Slot(name="openManagerWizard")
     def manage_plugins_wizard(self) -> None:
-        self.wizard = PluginsManagerWizard(self.window)
+        self.wizard = PluginsManagerWizard(application.main_window)
         self.wizard.show()
 
     def load_plugins(self):
@@ -68,7 +57,7 @@ class PluginController(QObject):
             plugin.load()
             # Add plugins tabs
             for tab in plugin.tabs:
-                self.window.add_tab_to_panel(tab, plugin.infos["name"], tab.panel)
+                application.main_window.add_tab_to_panel(tab, plugin.infos["name"], tab.panel)
             log.info("Loaded tab {}".format(name))
             return
         log.info("Removing plugin {}".format(name))
@@ -78,7 +67,7 @@ class PluginController(QObject):
         self.close_plugin_tabs(self.plugins[name])
 
     def close_plugin_tabs(self, plugin):
-        for panel in (self.window.left_panel, self.window.right_panel):
+        for panel in (application.main_window.left_panel, application.main_window.right_panel):
             panel.close_tab_by_tab_name(plugin.infos["name"])
 
     def reload_plugins(self):
@@ -93,8 +82,11 @@ class PluginController(QObject):
             except:
                 log.error(f"Error: plugin {name} not installed")
 
-    def close_plugins(self):
+    def close(self):
         """ close all plugins
         """
         for plugin in self.plugins.values():
             plugin.close()
+
+
+plugin_controller = PluginController(application)
