+++ activity_browser/ui/tables/models/base.py
@@ -2,6 +2,7 @@
 from typing import Optional
 
 import datetime
+
 import arrow
 import numpy as np
 import pandas as pd
@@ -10,14 +11,9 @@ from PySide2.QtCore import (
 )
 from PySide2.QtGui import QBrush
 
+from activity_browser import log
 from activity_browser.bwutils import commontasks as bc
 from activity_browser.ui.style import style_item
-
-import logging
-from activity_browser.logger import ABHandler
-
-logger = logging.getLogger('ab_logs')
-log = ABHandler.setup_with_logger(logger, __name__)
 
 
 class PandasModel(QAbstractTableModel):
@@ -370,11 +366,11 @@ class BaseTreeModel(QAbstractItemModel):
         """
         if item == None:
             return self.root
-        if item.childCount() > 0: #if its not a leaf
-            return item.child(0) # return the first child
-        if item.parent().childCount() > item.row()+1: #if there's still a sibling
+        if item.childCount() > 0:  # if its not a leaf
+            return item.child(0)  # return the first child
+        if item.parent().childCount() > item.row()+1:  # if there's still a sibling
             return item.parent().child(item.row()+1)
-        else: # look for siblings from previous "generations"
+        else:  # look for siblings from previous "generations"
             parent = item.parent()
             while parent != self.root:
                 if parent.parent().childCount() > parent.row() + 1:
@@ -413,6 +409,7 @@ class BaseTreeModel(QAbstractItemModel):
 
     def sync(self, *args, **kwargs) -> None:
         pass
+
 
 class ABSortProxyModel(QSortFilterProxyModel):
     """Reimplementation to allow for sorting on the actual data in cells instead of the visible data.
